requirements.txt
google-cloud-storage
python-multipart
scikit-learn
fastapi[all]
gunicorn
pytest
pandas
numpy
